requirements.txt
requests==2.27.1
Flask==2.0.2
flask-cors==3.0.10
python-dotenv==0.19.2
pyOpenSSL==22.0.0
Flask-Login==0.6.3
Werkzeug==2.2.2
secrets==1.0.2
PyJWT==2.8.0
pybase64==1.3.2
